Increased node-multiplication + added recentering
Increased node-multiplication of the lateral roots from 3 to 4, to
capture the lower growth rates.
Added recentering via a new function in the Root class, every node that
is checked is re-centered during each run (might need to extend this to
every node in general, as only few nodes are checked every round), to
correct for small changes in position.